reorganize type in chart package
