+++ packages/superset-ui-chart/src/query/buildQueryObject.ts
@@ -1,11 +1,9 @@
-import { FormData, getGranularity } from './FormData';
-import { Metric, Metrics } from './Metric';
+import Metrics from './Metrics';
+import { QueryObject } from '../types/Query';
+import { FormData } from '../types/FormData';
 
-// TODO: fill out the rest of the query object
-export interface QueryObject {
-  granularity: string;
-  groupby?: string[];
-  metrics?: Metric[];
+function getGranularity(formData: FormData): string {
+  return 'granularity_sqla' in formData ? formData.granularity_sqla : formData.granularity;
 }
 
 // Build the common segments of all query objects (e.g. the granularity field derived from
@@ -13,7 +11,7 @@ export interface QueryObject {
 // buildQuery method for each viz type (see `wordcloud/buildQuery.ts` for an example).
 // Note the type of the formData argument passed in here is the type of the formData for a
 // specific viz, which is a subtype of the generic formData shared among all viz types.
-export function buildQueryObject<T extends FormData>(formData: T): QueryObject {
+export default function buildQueryObject<T extends FormData>(formData: T): QueryObject {
   return {
     granularity: getGranularity(formData),
     metrics: new Metrics(formData).getMetrics(),
