+++ packages/superset-ui-chart/src/types/Annotation.ts
@@ -1,0 +1,6 @@
+/* eslint-disable import/prefer-default-export */
+
+export type AnnotationLayerMetadata = {
+  name: string;
+  sourceType?: string;
+};
