+++ packages/superset-ui-chart/src/types/Datasource.ts
@@ -1,0 +1,6 @@
+/* eslint-disable import/prefer-default-export */
+
+export enum DatasourceType {
+  Table = 'table',
+  Druid = 'druid',
+}
