+++ packages/superset-ui-chart/src/types/Query.ts
@@ -1,0 +1,26 @@
+import ChartProps from '../models/ChartProps';
+import { DatasourceType } from './Datasource';
+import { FormData } from './FormData';
+import { Metric } from './Metric';
+
+export interface QueryObject {
+  granularity: string;
+  groupby?: string[];
+  metrics?: Metric[];
+}
+
+export interface QueryContext {
+  datasource: {
+    id: number;
+    type: DatasourceType;
+  };
+  queries: Array<QueryObject>;
+}
+
+export type BuildQueryFunction<T extends FormData> = (formData: T) => QueryContext;
+
+export type TransformPropsFunction = (
+  chartProps: ChartProps,
+) => {
+  [key: string]: any;
+};
