+++ packages/superset-ui-core/src/types.ts
@@ -1,3 +1,5 @@
+/* eslint-disable import/prefer-default-export */
+
 export enum OverwritePolicy {
   ALLOW = 'ALLOW',
   PROHIBIT = 'PROHIBIT',
